converted the concurrency test to an automated test
